scripts/requirement.txt
Package             Version      Editable project location
------------------- ------------ -------------------------
addict              2.4.0
blis                0.7.9
catalogue           2.0.8
certifi             2022.12.7
charset-normalizer  3.1.0
click               8.1.3
coloredlogs         15.0.1
confection          0.0.4
contourpy           1.0.7
cycler              0.11.0
cymem               2.0.7
en-core-web-sm      3.5.0
filelock            3.11.0
flatbuffers         23.5.9
fonttools           4.39.3
huggingface-hub     0.13.4
humanfriendly       10.0
idna                3.4
importlib-resources 5.12.0
Jinja2              3.1.2
kiwisolver          1.4.4
langcodes           3.3.0
MarkupSafe          2.1.2
matplotlib          3.7.1
mmcls               0.25.0
mmcv-full           1.3.18
mmdet               2.25.3
mmsegmentation      0.26.0
mpmath              1.3.0
murmurhash          1.0.9
numpy               1.24.2
onnx                1.14.0
onnxruntime         1.14.1
opencv-python       4.7.0.72
packaging           23.0
pathy               0.10.1
Pillow              9.5.0
pip                 23.0.1
preshed             3.0.8
prettytable         3.7.0
protobuf            4.23.1
pycocotools         2.0.6
pydantic            1.10.7
pyparsing           3.0.9
python-dateutil     2.8.2
PyYAML              6.0
regex               2023.3.23
requests            2.28.2
scipy               1.10.1
segment-anything    1.0          /mnt/Git/segment-anything
setuptools          65.6.3
six                 1.16.0
smart-open          6.3.0
spacy               3.5.1
spacy-legacy        3.0.12
spacy-loggers       1.0.4
srsly               2.4.6
sympy               1.12
terminaltables      3.1.10
thinc               8.1.9
tokenizers          0.13.3
torch               1.9.1+cu111
torchaudio          0.9.1
torchvision         0.10.1+cu111
tqdm                4.65.0
transformers        4.27.4
typer               0.7.0
typing_extensions   4.5.0
urllib3             1.26.15
wasabi              1.1.1
wcwidth             0.2.6
wheel               0.38.4
yapf                0.32.0
zipp                3.15.0
